+++ dsfhub/resource_cloud_account.go
@@ -368,6 +368,11 @@ func resourceCloudAccount() *schema.Resource {
 				Description: "The jsonarUid unique identifier of the agentless gateway. Example: '7a4af7cf-4292-89d9-46ec-183756ksdjd'",
 				Required:    true,
 			},
+			"id": {
+				Type:        schema.TypeString,
+				Description: "Unique identifier for the asset",
+				Computed:    true,
+			},
 			"jsonar_uid": {
 				Type:        schema.TypeString,
 				Description: "Unique identifier (UID) attached to the Agentless Gateway controlling the asset",
@@ -553,6 +558,7 @@ func resourceCloudAccountReadContext(ctx context.Context, d *schema.ResourceData
 	}
 	d.Set("criticality", cloudAccountReadResponse.Data.AssetData.Criticality)
 	d.Set("gateway_id", cloudAccountReadResponse.Data.GatewayID)
+	d.Set("id", cloudAccountReadResponse.Data.ID)
 	d.Set("jsonar_uid", cloudAccountReadResponse.Data.AssetData.JsonarUID)
 	d.Set("location", cloudAccountReadResponse.Data.AssetData.Location)
 	d.Set("managed_by", cloudAccountReadResponse.Data.AssetData.ManagedBy)